feat: make project as package
