+++ .gitignore
@@ -1,22 +1,8 @@
-# See https://help.github.com/ignore-files/ for more about ignoring files.
-
-# dependencies
-/node_modules
-/.cahche
-
-# testing
-/coverage
-
-# production
-/build
-/dist
-
-# misc
 .DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
+
+/node_modules
+
+/dist
 
 npm-debug.log*
 yarn-debug.log*
