+++ .storybook/main.js
@@ -1,10 +1,3 @@
 module.exports = {
-  stories: ['../src/**/*.stories.(js|mdx)'],
-  addons: [
-    '@storybook/preset-create-react-app',
-    '@storybook/addon-actions',
-    '@storybook/addon-links',
-    '@storybook/addon-notes',
-    '@storybook/addon-docs'
-  ]
+  stories: ['../stories/**/*.stories.js']
 }
