+++ .storybook/preview-head.html
@@ -1,0 +1,2 @@
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/semantic-ui@2.4.2/dist/semantic.min.css">
+<link rel="stylesheet" href="index.css">
