+++ public/index.css
@@ -51,7 +51,6 @@
 }
 
 body {
-  margin: 0;
   font-family: Arial, Helvetica, sans-serif, -apple-system, 'Microsoft Yahei',
     'PingFangSC-Regular', 'Source Han Sans CN';
   -webkit-font-smoothing: antialiased;
