--- src/App.css
@@ -1,30 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 1vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.App-link:hover {
-  text-decoration: underline;
-}
-
-code {
-  font-style: italic;
-  font-weight: 600;
-  color: #61dafb;
-}
