--- src/App.test.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import { render } from '@testing-library/react'
-import App from './App'
-
-test('renders learn react link', () => {
-  const { getByText } = render(<App />)
-  const linkElement = getByText(/learn storybook/i)
-  expect(linkElement).not.toBe(null)
-  expect(linkElement).toMatchSnapshot()
-})
