--- src/components/index.ts
@@ -1,2 +1,0 @@
-export { default as TiButton } from './Button'
-export { default as TiCard } from './Card'
