+++ src/index.ts
@@ -1,0 +1,2 @@
+export { default as TiButton } from 'components/Button'
+export { default as TiCard } from 'components/Card'
