--- src/stories/0-Welcome.stories.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { linkTo } from '@storybook/addon-links'
-import { Welcome } from '@storybook/react/demo'
-
-export default {
-  title: 'Welcome',
-  component: Welcome
-}
-
-export const ToStorybook = () => <Welcome showApp={linkTo('Button')} />
-
-ToStorybook.story = {
-  name: 'to Storybook'
-}
